move string length checks and truncation to module
