+++ context/src/junit/validator.rs
@@ -2,7 +2,7 @@ use chrono::{DateTime, FixedOffset, Utc};
 use quick_junit::Report;
 use thiserror::Error;
 
-use crate::{validate_field_len, FieldLen};
+use crate::string_safety::{validate_field_len, FieldLen};
 
 use super::parser::extra_attrs;
